moving dilation and erode to filters
