+++ src/porespy/filters/__init__.py
@@ -17,29 +17,36 @@ image, but with altered values.
     filters.apply_chords
     filters.apply_chords_3D
     filters.apply_padded
+    filters.capillary_transform
     filters.chunked_func
+    filters.dilate
     filters.distance_transform_lin
+    filters.erode
     filters.fftmorphology
     filters.fill_closed_pores
+    filters.fill_trapped_voxels
+    filters.find_closed_pores
     filters.find_disconnected_voxels
     filters.find_dt_artifacts
+    filters.find_invalid_pores
     filters.find_peaks
+    filters.find_surface_pores
     filters.find_trapped_regions
     filters.flood
     filters.flood_func
     filters.hold_peaks
-    filters.ibip
-    filters.ibip_gpu
     filters.local_thickness
     filters.nl_means_layered
     filters.nphase_border
     filters.pc_to_satn
+    filters.pc_to_seq
     filters.porosimetry
     filters.prune_branches
     filters.reduce_peaks
     filters.region_size
     filters.satn_to_seq
     filters.seq_to_satn
+    filters.size_to_pc
     filters.size_to_satn
     filters.size_to_seq
     filters.snow_partitioning
@@ -51,6 +58,7 @@ image, but with altered values.
     filters.trim_nearby_peaks
     filters.trim_nonpercolating_paths
     filters.trim_saddle_points
+    filters.trim_saddle_points_legacy
     filters.trim_small_clusters
 
 """
@@ -63,3 +71,4 @@ from ._size_seq_satn import *
 from ._snows import *
 from ._transforms import *
 from ._invasion import *
+from ._morphology import *
